% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{add_estMeanStructure}
\alias{add_estMeanStructure}
\title{Internal helper function for eval_meanStructure}
\usage{
add_estMeanStructure(dat_trueMean, reg_obj, srvf_regularGrid)
}
\arguments{
\item{dat_trueMean}{Dataset containing the simulated mean structure.}

\item{reg_obj}{Registration object, as returned by \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}.}

\item{srvf_regularGrid}{Vector of the regular time grid. Only needed if
\code{reg_obj} was estimated with \code{align_fPCA}.}
}
\value{
\code{dat_trueMean} with additional columns, see above description.
}
\description{
Internal helper function to be called inside \code{\link{eval_meanStructure}},
for the comparison of a curve's mean structure
(global mean + curve-specific functional principal component structure) with
its true (simulated) mean structure.
This function takes a dataset of the true (simulated) mean structure with
columns \code{"index_internal"} and \code{"value"} (both containing the
underlying mean structure) as well as \code{"index_observed"} (containing the
warped version of \code{"index_internal"} based on the simulated warping).
Then, based on an object estimated either with \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}, the function adds two new columns to
this dataset: \cr
\enumerate{
\item \code{"index_internal_est"}: These are the index values on the registered
time domain that correspond to 'index_observed' based on the estimated
warping functions. \cr
\item \code{"value_est"}: These are the values of the estimated mean structure
(global mean + curve-specific FPC structure) evaluated at
\code{"index_internal_est"}. \cr
With this information \code{"value"} and \code{"value_est"} can be directly
compared since both refer to the (true and estimated) mean structures
evaluated on the observed domain. \cr
Note: If \code{reg_obj} was computed with \code{align_fPCA},
the \code{"value_est"} values are computed by first extracting the estimated
mean structure from the \code{reg_obj} object and then linearly
interpolating it.
}
}
