% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{calcMSE_domainLengths}
\alias{calcMSE_domainLengths}
\title{Evaluation of the estimated domain lengths}
\usage{
calcMSE_domainLengths(dat_trueDomainLengths, dat_estDomainLengths)
}
\arguments{
\item{dat_trueDomainLengths}{Dataset of simulated curves, with relevant columns
\code{c("id","index","value")}.}

\item{dat_estDomainLengths}{Dataset of registered curves, with relevant columns
\code{c("id","t_hat","value")}. E.g. the object \code{obj$Y} where \code{obj}
was returned by \code{\link[registr]{register_fpca}}.}
}
\value{
Calculated MSE.
}
\description{
Function to evaluate the performance in estimating phase variation by
calculating the mean squared error (MSE) between the simulated and estimated
domain lengths of the individual curves.
}
