% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{calcMSE_globalMean}
\alias{calcMSE_globalMean}
\title{Internal helper function for eval_globalMean}
\usage{
calcMSE_globalMean(dat_trueMean, dat_estMean)
}
\arguments{
\item{dat_trueMean, dat_estMean}{Datasets containing the simulated and estimated
global mean curve, respectively, with relevant columns \code{"index_internal"}
and \code{"value"}.}
}
\value{
MSE value.
}
\description{
Internal helper function to be called from within \code{\link{eval_globalMean}},
to calculate the MSE between the simulated and estimated global mean curve.
The comparison is based on the internal (registered) time domain.
}
\details{
Since the index values are never exactly identical between \code{dat_trueMean}
and \code{dat_estMean}, the latter is linearly interpolated.
}
