% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{calcWeightedMISE_GFPCA}
\alias{calcWeightedMISE_GFPCA}
\title{Internal helper function for eval_FPCstructure}
\usage{
calcWeightedMISE_GFPCA(
  dat_trueFPCs,
  dat_estFPCs,
  eigenvalues_sim,
  return_datasets = FALSE
)
}
\arguments{
\item{dat_trueFPCs, dat_estFPCs}{Datasets of true and estimated FPCs, with columns
\code{c("id","index","value")}.}

\item{eigenvalues_sim}{Numeric vector of simulated eigenvalues. Their square
roots are used as weights for the wMISE calculation.}

\item{return_datasets}{Indicator if datasets for the true and estimated
FPCs (the latter scaled with -1, if necessary) should be returned
additionally to the wMISE. Defaults to \code{FALSE}.}
}
\value{
Numeric wMISE value. If \code{return_datasets == TRUE} returns a list
which additionally contains datasets containing the true and estimated FPCs.
}
\description{
Helper function to be called internally from \code{\link{eval_FPCstructure}},
to calculate the goodness-of-fit for a GFPCA solution as the weighted MISE
over all functional principal components (FPCs) between the true and the
estimated FPC values. The square roots of the true, simulated eigenvalues are
used as weights. \cr
For each FPC, its ISE value is calculated by taking the minimum of
\code{c(<ISE to +1*FPC>, <ISE to -1*FPC>)}.
}
