% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{eval_FPCstructure}
\alias{eval_FPCstructure}
\title{Evaluation of the estimated FPC structure based on wMISE}
\usage{
eval_FPCstructure(
  FPCdat_sim,
  reg_obj,
  eigenvalues_sim,
  srvf_regularGrid,
  return_datasets = FALSE
)
}
\arguments{
\item{FPCdat_sim}{Dataset containing the simulated FPCs, with columns
\code{c("id","index","value")}.}

\item{reg_obj}{Registration object, as returned by \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}.}

\item{eigenvalues_sim}{Numeric vector of simulated eigenvalues. Their square
roots are used as weights for the wMISE calculation.}

\item{srvf_regularGrid}{Vector of the regular time grid. Only needed if
\code{reg_obj} was estimated with \code{align_fPCA}.}

\item{return_datasets}{Indicator if datasets for the true and estimated
FPCs (the latter scaled with -1, if necessary) should be returned
additionally to the wMISE. Defaults to \code{FALSE}.}
}
\value{
List of wMISE values, based on (i) all FPCs, (ii) only the first FPC,
(iii) only the first two FPCs. If \code{return_datasets == TRUE} the returned
list also contains two datasets containing the true and estimated versions of
the FPCs.
}
\description{
Function to evaluate the performance in estimating the functional principal
component (FPC) structure by calculating the weighted mean integrated squared
error (wMISE) between the simulated and estimated FPCs. The weighted arithmetic
mean is taken of the FPC-specific ISE values by weighting them with the
square roots of the simulated eigenvalues. \cr
Note that this function is also applicable to models estimated with the SRVF
approach. However, this in not used in the paper since the SRVF approach
only estimates orthonormal FPCs in the SRVF space, but not in the original
function space.
}
\details{
For each FPC, its ISE value is calculated by taking the minimum of
\code{c(<ISE to +1*FPC>, <ISE to -1*FPC>)}.

Some notes regarding the output of \code{\link[fdasrvf]{align_fPCA}}: \cr
\enumerate{
\item The FPCs are available in \code{x$vfpca$U}, the scores in \code{x$vfpca$coef} \cr
\item Weirdly, nrow of the eigenfunction matrix doesn't equal \code{length(regular_grid)},
but \code{length(regular_grid) + 1}... We simply use the rows \code{1:(nrow - 1)}
of the matrix to have something matching with \code{length(regular_grid)}.
}
}
