% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{eval_FPCstructure_LV}
\alias{eval_FPCstructure_LV}
\title{Evaluation of the estimated FPC structure based on the LV measure}
\usage{
eval_FPCstructure_LV(FPCdat_sim, reg_obj, srvf_regularGrid)
}
\arguments{
\item{FPCdat_sim}{Dataset containing the simulated FPCs, with columns
\code{c("id","index","value")}.}

\item{reg_obj}{Registration object, as returned by \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}.}

\item{srvf_regularGrid}{Vector of the regular time grid. Only needed if
\code{reg_obj} was estimated with \code{align_fPCA}.}
}
\value{
LV measure.
}
\description{
Function to evaluate the performance in estimating the functional principal
component (FPC) structure by calculating the adapted version of the
Larsson-Villani measure as introduced in the paper to compare not the
individual FPCs, but the span of the overall FPC basis. \cr
Note that this function is also applicable to models estimated with the SRVF
approach. However, this in not used in the paper since the SRVF approach
only estimates orthonormal FPCs in the SRVF space, but not in the original
function space.
}
\details{
Some notes regarding the output of \code{\link[fdasrvf]{align_fPCA}}: \cr
\enumerate{
\item The FPCs are available in \code{x$vfpca$U}, the scores in \code{x$vfpca$coef} \cr
\item Weirdly, nrow of the eigenfunction matrix doesn't equal \code{length(regular_grid)},
but \code{length(regular_grid) + 1}... We simply use the rows \code{1:(nrow - 1)}
of the matrix to have something matching with \code{length(regular_grid)}.
}
}
