% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{eval_globalMean}
\alias{eval_globalMean}
\title{Evaluation of the estimated global mean}
\usage{
eval_globalMean(
  dat_sim,
  dat_simRegular,
  distribution,
  reg_obj,
  registr_fpcaType,
  return_datasets = FALSE
)
}
\arguments{
\item{dat_sim}{Simulated dataset.}

\item{dat_simRegular}{Only needed if \code{reg_obj} was estimated with
\code{align_fPCA}. Interpolated version of \code{dat_sim} s.t. all curves are
observed on a regular grid.}

\item{distribution}{One of \code{c("gaussian","gamma")}, specifying if the
model was estimated with Gamma family.}

\item{reg_obj}{Registration object, as returned by \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}.}

\item{registr_fpcaType}{One of \code{c("variationalEM","two-step")}. Only
needed if \code{reg_obj} was estimated with \code{register_fpca}.}

\item{return_datasets}{Indicator if datasets for the true and estimated versions
of the global mean curve should be returned additionally to
the MSE value. Defaults to \code{FALSE}.}
}
\value{
If \code{return_datasets == FALSE} only returns the calculated MSE
value. If \code{return_datasets == TRUE} returns a list with the MSE value
and two datasets containing the true and estimated versions of the
global mean curve.
}
\description{
Function to evaluate the performance in estimating the global mean curve by
calculating the mean squared error (MSE) between the simulated and the
estimated global mean.
}
