% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{eval_meanStructure}
\alias{eval_meanStructure}
\title{Evaluation of the curve-specific mean structure}
\usage{
eval_meanStructure(
  dat_sim,
  reg_obj,
  registr_fpcaType,
  srvf_regularGrid,
  return_datasets = FALSE
)
}
\arguments{
\item{dat_sim}{Simulated dataset.}

\item{reg_obj}{Registration object, as returned by \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}.}

\item{registr_fpcaType}{One of \code{c("variationalEM","two-step")}. Only
needed if \code{reg_obj} was estimated with \code{register_fpca}.}

\item{srvf_regularGrid}{Vector of the regular time grid. Only needed if
\code{reg_obj} was estimated with \code{align_fPCA}.}

\item{return_datasets}{Indicator if datasets for the true and estimated versions
of the (represented) curves without noise should be returned additionally to
the MISE value. Defaults to \code{FALSE}.}
}
\value{
If \code{return_datasets == FALSE} only returns the calculated MISE
value. If \code{return_datasets == TRUE} returns a list with the MISE value
and two datasets containing the true and estimated versions of the
(represented) curves without noise.
}
\description{
Function to evaluate the curve-specific mean structure by calculating the
mean integrated squared error (MISE) between (a) the simulated curves
(before adding random noise) and (b) the represented curves by the estimated
solution (i.e. estimated global mean structure + curve-specific functional
principal component structure).
}
