% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{eval_warpingFunctions}
\alias{eval_warpingFunctions}
\title{Evaluation of the estimated warping functions}
\usage{
eval_warpingFunctions(dat_sim, reg_obj, srvf_regularGrid)
}
\arguments{
\item{dat_sim}{Simulated dataset.}

\item{reg_obj}{Registration object, as returned by \code{\link[registr]{register_fpca}}
or \code{\link[fdasrvf]{align_fPCA}}.}

\item{srvf_regularGrid}{Vector of the regular time grid. Only needed if
\code{reg_obj} was estimated with \code{align_fPCA}.}
}
\value{
Calculated MISE.
}
\description{
Function to evaluate the performance in estimating phase variation by
calculating the mean integrated squared error (MISE) between the simulated
and estimated warping functions.
}
\details{
If \code{reg_obj} was estimated with \code{\link[fdasrvf]{align_fPCA}} the
simulated warping functions are first linearly interpolated to the same
regular grid \code{srvf_regularGrid} that was used in the estimation step.
}
