% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_algorithmWrappers.R
\name{FGAMM_wrapper}
\alias{FGAMM_wrapper}
\title{Wrapper to call FGAMM with assumed completeness}
\usage{
FGAMM_wrapper(data, job, instance, max_iterations, n_cores)
}
\arguments{
\item{data}{General arguments as requested by \code{batchtools}.}

\item{job}{General arguments as requested by \code{batchtools}.}

\item{instance}{A specific simulated data setting as returned by
\code{\link{simData_wrapper}}.}

\item{max_iterations}{Maximum number of iterations of the joint algorithm
of \code{register_fpca}.}

\item{n_cores}{Number of cores to use for computation.}
}
\value{
List with performance criteria.
}
\description{
Wrapper function to call FGAMM with Gaussian or Gamma family and assumed
completeness in the simulation study. Written specifically for the use with
the \code{batchtools} package.
}
