% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_algorithmWrappers.R
\name{FGAMMWithIncompleteness_wrapper}
\alias{FGAMMWithIncompleteness_wrapper}
\title{Wrapper to call FGAMM with potentially assumed incompleteness}
\usage{
FGAMMWithIncompleteness_wrapper(
  data,
  job,
  instance,
  lambda_inc = 0,
  max_iterations,
  n_cores,
  incompleteness = TRUE
)
}
\arguments{
\item{data, job}{General arguments as requested by \code{batchtools}.}

\item{instance}{A specific simulated data setting as returned by
\code{\link{simData_wrapper}}.}

\item{lambda_inc}{Penalization parameter passed to \code{\link[registr]{register_fpca}}.
Defaults to 0.}

\item{max_iterations}{Maximum number of iterations of the joint algorithm
of \code{register_fpca}.}

\item{n_cores}{Number of cores to use for computation.}

\item{incompleteness}{Indicator if incompleteness should be assumed or not.
Defaults to \code{TRUE}.}
}
\value{
List with performance criteria.
}
\description{
Wrapper function to call FGAMM with Gaussian or Gamma family in the simulation
study. Written specifically for the use with the \code{batchtools} package.
}
