% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RegResults.R
\name{get_dataList_representedCurves}
\alias{get_dataList_representedCurves}
\title{Retrieve datasets of the represented curves and the mean curve}
\usage{
get_dataList_representedCurves(fpca_obj, exp_transformation = FALSE)
}
\arguments{
\item{fpca_obj}{Element \code{fpca_obj} of an object estimated with
\code{\link[registr]{register_fpca}}.}

\item{exp_transformation}{Indicator if a log link was used and data should
be exp transformed. Defaults to \code{FALSE}.}
}
\value{
List of two \code{data.frame}s for (i) the represented curves and
(ii) the estimated mean curve.
}
\description{
Internal helper function based on the output of \code{\link[registr]{register_fpca}}.
The function returns a list with (i) a dataset with the represented version
of each observed curve (i.e. mean + individual FPC structure), and (ii) a
dataset of the estimated mean curve.
}
