% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seismicApplication.R
\name{get_finalCircleDat}
\alias{get_finalCircleDat}
\title{Create a dataset of radial auxiliary lines with different radii}
\usage{
get_finalCircleDat(dat)
}
\arguments{
\item{dat}{Seismic dataset with columns \code{c("id","receiver.x","receiver.y")},
where the latter define the Easting and Northing coordinate of the virtual
seismometers in meters. The dataset must contain information on the seismometer
which is located directly at the epicenter (i.e. curve id's ending on \code{"_62"}).}
}
\value{
\code{data.frame} of points on the different radial auxiliary lines.
}
\description{
This is a wrapper function which calls \code{\link{get_circle_points}}
multiple times to create a dataset for plotting radial auxiliary lines
with different radii specifically for the seismic dataset.
Values 5km, 10km, 15km, 20km, 25km are used as radii.¸
}
