% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_evalPerformance.R
\name{lv_overlap}
\alias{lv_overlap}
\title{Calculate the adapted version of the measure of Larsson and Villani (2001)}
\usage{
lv_overlap(a, b, dim_refBase = min(NCOL(a), NCOL(b)))
}
\arguments{
\item{a, b}{Matrices with the FPCs as columns, evaluated on the same index grid.}

\item{dim_refBase}{Dimension of the reference basis. E.g., if \code{a} contains
the simulated FPC basis and \code{b} the estimated basis, \code{dim_refBase}
should be set to \code{ncol(a)}. Defaults to \code{min(NCOL(a), NCOL(b))}.}
}
\value{
Calculated LV measure.
}
\description{
Function to calculate the adaption of the measure of Larsson and Villani (2001)
as outlined in the paper, to calculate the overlap between a simulated and
an estimated functional principal component (FPC) basis. The resulting measure
is between 0 and 1.
}
