% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seismicApplication.R
\name{plot_ampVar}
\alias{plot_ampVar}
\title{Plot the estimated amplitude variation vs hypocentral distance and the dynamic friction}
\usage{
plot_ampVar(dat_scores, FPC_index, hide_yLabels = FALSE, base_size = 16)
}
\arguments{
\item{dat_scores}{\code{data.frame} with one row per observed curve and
columns containing the hypocentral distance of the virtual seismometer in
meters (column \code{"hypo.dis"}), the Easting and Northing coordinate of
the seismometer in meters (\code{"receiver.x"} and \code{"receiver.y"}),
the dynamic coefficient of friction of the seismic simulation (\code{"md"})
and one column \code{"score_FPCx"} for each estimated FPC \code{"x"}.}

\item{FPC_index}{Numeric index specifying which FPC should be plotted.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Hexagonal binning plot function to visualize the estimated scores for a
specific functional principal component (FPC) versus the hypocentral distance
of the virtual seismogram and the dynamic coefficient of friction of the
seismic simulation.
}
