% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RegResults.R
\name{plot_domainDilation}
\alias{plot_domainDilation}
\title{Visualize the estimated domain dilation}
\usage{
plot_domainDilation(Y, base_size = 16, hide_yLabels = FALSE)
}
\arguments{
\item{Y}{Element \code{obj$Y} where \code{obj} is an object as returned
by \code{\link[registr]{register_fpca}}.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Plot the overall domain dilation of a model estimated with
\code{registr::register_fpca}.
}
