% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RegResults.R
\name{plot_FPC}
\alias{plot_FPC}
\title{Plot one estimated functional principal component}
\usage{
plot_FPC(
  reg,
  FPC_index,
  title = "<my_title>",
  sd_factor = 2,
  log10_scale = FALSE,
  ylim = NULL,
  ylab = "",
  base_size = 16,
  xlab = "t [registered]",
  hide_xLabels = FALSE,
  hide_yLabels = FALSE
)
}
\arguments{
\item{reg}{Object estimated with \code{\link[registr]{register_fpca}}.}

\item{FPC_index}{Numeric index specifying which FPC should be plotted.}

\item{title}{Character title of the plot. Pass \code{NULL} to suppress the title.}

\item{sd_factor}{Numeric passed to \code{registr:::plot.fpca}.}

\item{log10_scale}{Indicator if the curves should be plotted on log10 scale.
Defaults to \code{FALSE}.}

\item{ylim}{Optional limits for the y-axis.}

\item{ylab}{Optional labels for the x-axis and y-axis.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{xlab}{Optional labels for the x-axis and y-axis.}

\item{hide_xLabels}{Indicator if the labels on the x-axis should be suppressed.
Defaults to \code{FALSE}.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Plot a specific functional principal component (FPC) of a model estimated with
\code{registr::register_fpca}.
}
