% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_general.R
\name{plot_lasagna}
\alias{plot_lasagna}
\title{Lasagna plots of observed / registered curves}
\usage{
plot_lasagna(
  Y,
  title,
  index_var = "index",
  log10_scale = FALSE,
  xlim = NULL,
  xlab = index_var,
  ylab = "id",
  base_size = 16,
  alpha = 0.2,
  hide_yLabels = FALSE,
  legend.position = "bottom",
  legend_limits = NULL,
  legend_title = "value",
  line_size = 0.5
)
}
\arguments{
\item{Y}{\code{data.frame} with columns \code{c("id","index","value")}.
If the curves should be plotted over another domain as \code{"index"} this
can be specified by specifying argument \code{index_var}.}

\item{title}{Character title of the plot. Pass \code{NULL} to suppress the title.}

\item{index_var}{Character name of the index variable in \code{Y} which
specifies the domain over which the curves should be plotted. Defaults to
\code{"index"}.}

\item{log10_scale}{Indicator if the curves should be plotted on log10 scale.
Defaults to \code{FALSE}.}

\item{xlim}{Option limits for the x-axis.}

\item{xlab}{Optional labels for the x-axis and y-axis.}

\item{ylab}{Optional labels for the x-axis and y-axis.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{alpha}{Alpha value defining the transparency of the individual lines.
Defaults to 0.2.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}

\item{legend.position}{Character defining the position of the legend, as
passed to \code{\link[ggplot2]{theme}}. Defaults to \code{"bottom"}.}

\item{legend_limits}{Optional limits for the color scale.}

\item{legend_title}{Character title for the legend. Pass \code{NULL} to
suppress the title.}

\item{line_size}{Line size passed to \code{\link[ggplot2]{geom_line}}.
Defaults to 0.5.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Create a lasagna plot of observed or registered curves. The plot is
sorted by \code{levels(Y$id)} along the y-axis.
}
