% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seismicApplication.R
\name{plot_phaseVar}
\alias{plot_phaseVar}
\title{Plot the estimated phase variation vs hypocentral distance and the dynamic friction}
\usage{
plot_phaseVar(dat, index_ref = NA, hide_yLabels = FALSE, base_size = 16)
}
\arguments{
\item{dat}{\code{data.frame} of observed curves with columns
\code{c("id","index","value")} and with an additional column \code{"t_hat"}
which specifies the estimated registered time domain.}

\item{index_ref}{index value at which the amount of overall domain dilation
is looked at. If it's specified as \code{NA} (default), the overall domain
dilations are used of the individual curves.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Hexagonal binning plot function to visualize the estimated phase variation
versus the hypocentral distance of the virtual seismogram and the dynamic
coefficient of friction of the seismic simulation. Phase variation is
measured as the overall absolute domain dilation estimated for the first
\code{index_ref} seconds.
}
