% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seismicApplication.R
\name{plot_phaseVar_overSpace}
\alias{plot_phaseVar_overSpace}
\title{Plot the estimated phase variation over space}
\usage{
plot_phaseVar_overSpace(
  dat,
  index_ref = NA,
  plot_auxiliaryLines = TRUE,
  hide_yLabels = FALSE,
  base_size = 16
)
}
\arguments{
\item{dat}{\code{data.frame} of observed curves with columns
\code{c("id","index","value")} and with an additional column \code{"t_hat"}
which specifies the estimated registered time domain.}

\item{index_ref}{index value at which the amount of overall domain dilation
is looked at. If it's specified as \code{NA} (default), the overall domain
dilations are used of the individual curves.}

\item{plot_auxiliaryLines}{Indicator if radial auxiliary lines should be
drawn on the plot for highlighting 5km wide steps from the epicenter.
Defaults to \code{TRUE}.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Hexagonal binning plot function to visualize the estimated phase variation
over space. Phase variation is measured as the overall absolute domain
dilation estimated for the first \code{index_ref} seconds.
}
