% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RegResults.R
\name{plot_representedCurves_lasagna}
\alias{plot_representedCurves_lasagna}
\title{Lasagna plot of represented curves}
\usage{
plot_representedCurves_lasagna(
  fpca_obj,
  ids_sorted = NULL,
  exp_transformation = FALSE,
  log10_scale = FALSE,
  title = "<my_title>",
  base_size = 16,
  hide_yLabels = FALSE,
  ...
)
}
\arguments{
\item{fpca_obj}{Element \code{fpca_obj} of an object estimated with
\code{\link[registr]{register_fpca}}.}

\item{ids_sorted}{Optional vector of the ids in the data as they should be
sorted along the y-axis of the plot.}

\item{exp_transformation}{Indicator if a log link was used and data should
be exp transformed. Defaults to \code{FALSE}.}

\item{log10_scale}{Indicator if the curves should be plotted on log10 scale.
Defaults to \code{FALSE}.}

\item{title}{Character title of the plot. Pass \code{NULL} to suppress the title.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link{plot_lasagna}}.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Create a lasagna plot of the represented curves (overall mean + individual
FPC structure) based on a model estimated with \code{\link[registr]{register_fpca}},
including the estimated overall mean curve. By default, the plot is
sorted by \code{levels(fpca_obj$Yhat$id)} along the y-axis.
}
