% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RegResults.R
\name{plot_representedCurves_spaghetti}
\alias{plot_representedCurves_spaghetti}
\title{Spaghetti plot of represented curves}
\usage{
plot_representedCurves_spaghetti(
  fpca_obj,
  exp_transformation = FALSE,
  log10_scale = FALSE,
  title = "<my_title>",
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  y_breaks = waiver(),
  y_minorBreaks = waiver(),
  base_size = 16,
  alpha = 0.2,
  hide_yLabels = FALSE
)
}
\arguments{
\item{fpca_obj}{Element \code{fpca_obj} of an object estimated with
\code{\link[registr]{register_fpca}}.}

\item{exp_transformation}{Indicator if a log link was used and data should
be exp transformed. Defaults to \code{FALSE}.}

\item{log10_scale}{Indicator if the curves should be plotted on log10 scale.
Defaults to \code{FALSE}.}

\item{title}{Character title of the plot. Pass \code{NULL} to suppress the title.}

\item{xlab}{Optional labels for the x-axis and y-axis.}

\item{ylab}{Optional labels for the x-axis and y-axis.}

\item{ylim}{Optional limits for the y-axis.}

\item{y_breaks}{Optional breaks for the y-axis.}

\item{y_minorBreaks}{Optional breaks for the y-axis.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{alpha}{Alpha value defining the transparency of the individual lines.
Defaults to 0.2.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Create a spaghetti plot of the represented curves (overall mean + individual
FPC structure) based on a model estimated with \code{\link[registr]{register_fpca}},
including the estimated overall mean curve.
}
