% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simStudy.R
\name{plot_results_simStudy}
\alias{plot_results_simStudy}
\title{Plot the results of the simulation study}
\usage{
plot_results_simStudy(
  dat_pars,
  dat_res,
  corr_structure,
  log_scale = FALSE,
  show_allMeasures = FALSE,
  base_size = 12
)
}
\arguments{
\item{dat_pars, dat_res}{Main results tables from the simulation study.}

\item{corr_structure}{One of \code{c("ID","AP","AI")}.}

\item{log_scale}{Indicator if all y-axes should be log transformed. Defaults
to \code{FALSE}.}

\item{show_allMeasures}{Indicator if all measures should be plotted (for the
appendix plots) or not. Defaults to \code{FALSE}.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 12.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Plot function for the results table of the simulation study.
}
