% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_general.R
\name{plot_spaghetti}
\alias{plot_spaghetti}
\title{Spaghetti plots of observed / registered curves}
\usage{
plot_spaghetti(
  Y,
  title = "<my_title>",
  index_var = "index",
  log10_scale = FALSE,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  y_breaks = waiver(),
  y_minorBreaks = waiver(),
  base_size = 16,
  alpha = 0.2,
  hide_yLabels = FALSE
)
}
\arguments{
\item{Y}{\code{data.frame} with columns \code{c("id","index","value")}.
If the curves should be plotted over another domain as \code{"index"} this
can be specified by specifying argument \code{index_var}.}

\item{title}{Character title of the plot. Pass \code{NULL} to suppress the title.}

\item{index_var}{Character name of the index variable in \code{Y} which
specifies the domain over which the curves should be plotted. Defaults to
\code{"index"}.}

\item{log10_scale}{Indicator if the curves should be plotted on log10 scale.
Defaults to \code{FALSE}.}

\item{xlab, ylab}{Optional labels for the x-axis and y-axis.}

\item{ylim}{Optional limits for the y-axis.}

\item{y_breaks, y_minorBreaks}{Optional breaks for the y-axis.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{alpha}{Alpha value defining the transparency of the individual lines.
Defaults to 0.2.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Create a spaghetti plot of observed or registered curves.
}
