% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_general.R
\name{plot_template}
\alias{plot_template}
\title{Plot a template function}
\usage{
plot_template(
  dat,
  dat_template,
  title = "<my_title>",
  ylab = "Derivative",
  base_size = 16
)
}
\arguments{
\item{dat, dat_template}{\code{data.frame}s in the form of arguments \code{Y}
and \code{Y_template} as requested by function \code{\link[registr]{register_fpca}}.}

\item{title}{Character title of the plot. Pass \code{NULL} to suppress the title.}

\item{ylab}{Optional labels for the x-axis and y-axis.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Plot the template function used for the initial registration step in
\code{\link[registr]{register_fpca}} together with a spaghetti plot of the
raw data.
}
