% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RegResults.R
\name{plot_warpings}
\alias{plot_warpings}
\title{Plot inverse warping functions}
\usage{
plot_warpings(
  Y,
  ylab = "t [registered]",
  title = NULL,
  alpha = 0.2,
  base_size = 16,
  hide_yLabels = FALSE
)
}
\arguments{
\item{Y}{Element \code{obj$Y} where \code{obj} is an object as returned
by \code{\link[registr]{register_fpca}}.}

\item{ylab}{Character title for the y-axis. Defaults to \code{"t [registered]"}.}

\item{title}{Optional character title for the plot.}

\item{alpha}{Alpha value defining the transparency of the individual lines.
Defaults to 0.2.}

\item{base_size}{Base text size as passed to \code{\link[ggplot2]{theme_minimal}}.
Defaults to 16.}

\item{hide_yLabels}{Indicator if the labels on the y-axis should be suppressed.
Defaults to \code{FALSE}.}
}
\value{
Returns the plot in form of a \code{ggplot2} object.
}
\description{
Plot the inverse warping functions of a model estimated with
\code{registr::register_fpca}.
}
