% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_simCurvesWrapper.R
\name{simData_wrapper}
\alias{simData_wrapper}
\title{Wrapper for 'simulate_curves' for the simulation study}
\usage{
simData_wrapper(
  data,
  job,
  distribution,
  amplitude_rank,
  incompleteness_strength,
  correlation_structure,
  N,
  n_timeGrid
)
}
\arguments{
\item{data, job}{General arguments as requested by \code{batchtools}.}

\item{distribution}{One of \code{c("gaussian","gamma")}, specifying the
distribution of the data.}

\item{amplitude_rank}{One of \code{c("Rank 1","Rank 2-3","Rank 3-4")}.}

\item{incompleteness_strength}{One of \code{c("CC","WIC","SIC")},
specifying if complete curves (\code{"CC"}) should be simulated or curves
with weak incompleteness (\code{"WIC"}) or strong incompleteness
(\code{"SIC"}).}

\item{correlation_structure}{One of \code{c("ID","AP","AI","PI")}, specifying
if amplitude, phase, and the level of incompleteness should be mutually
independent (\code{"ID"}), if amplitude and phase should be correlated
(\code{"AP"}),if amplitude and the level of incompleteness should be
correlated (\code{"AI"}), or if phase and the level of incompleteness should
be correlated (\code{"PI"}).}

\item{N}{Number of curves to simulate.}

\item{n_timeGrid}{Number of (regular) evaluation points over the complete
time grid.}
}
\value{
List with (i) the simulated data, (ii) a dataset containing the
underlying functional principal components, (iii) a list with the main
settings used for the data simulation.
}
\description{
High-level wrapper function for \code{\link{simulate_curves}} for
simulating curves in the simulation study. Written specifically for the
use with the \code{batchtools} package.
}
