% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_curves.R
\name{simulate_curves}
\alias{simulate_curves}
\title{Simulate (in)complete curves with random warping and an FPC structure}
\usage{
simulate_curves(
  N = 20,
  n_timeGrid = 50,
  distribution = "gaussian",
  tDistribution_df = 3,
  random_warping = FALSE,
  incompleteness = FALSE,
  incompleteness_rate = 0.5,
  FPCA_structure = NULL,
  corr_amplitude_phase = 0,
  corr_amplitude_incompleteness = 0,
  corr_phase_incompleteness = 0,
  seed = NULL
)
}
\arguments{
\item{N}{Number of curves to simulate. Defaults to 20.}

\item{n_timeGrid}{Number of (regular) evaluation points over the complete
time grid. Defaults to 50.}

\item{distribution}{One of \code{c("gaussian","t","gamma")}, specifying the
distribution of the data. Defaults to \code{"gaussian"}.}

\item{tDistribution_df}{If \code{distribution = "t"}, the degrees of freedom
for the t distributed white noise simulated with \code{\link{rt}}. Defaults
to 3.}

\item{random_warping}{Indicator if the regular time grid should be randomly
warped. Defaults to \code{FALSE}.}

\item{incompleteness, incompleteness_rate}{Indicator if simulated curves
should be incomplete. Incompleteness is simulated by sampling individual
random cut-off indices in the latter part of the domain. This 'latter part'
are the last \code{<incompleteness_rate * 100>\%} of the domain.
Defaults to \code{FALSE} and the last 50\code{\%} of the domain.}

\item{FPCA_structure}{Optional named list with elements \code{"n_FPCs"}
(an integer between 1 and 4) and \code{"eigenvalues"} (a non-negative numeric
vector of length \code{n_FPCs} that sums to 1. The square roots are used as standard deviation
of a normal distribution to sample the individual FPC scores.). If given, structural
variation is added around the mean based on given functional principal
components.}

\item{corr_amplitude_phase}{Optional correlation between amplitude variation
(specifically the maximum peak) and phase variation (specifically how much
the initial part of the time domain is expanded). This parameter doesn't lead
to the two measures having exactly the specified correlation, but should be
seen as a general parameter, higher (absolute) values of which lead to
a clearer association. Can have values between -1 and 1. Defaults to 0.}

\item{corr_amplitude_incompleteness}{Optional correlation between amplitude
variation (specifically the maximum peak) and the curve's incompleteness
(specifically the length of the curve). This parameter doesn't lead to the
two measures having exactly the specified correlation, but should be seen as
a general parameter, higher (absolute) values of which lead to a clearer
association. Can have values between -1 and 1. Defaults to 0. Only one of
\code{corr_amplitude_incompleteness} / \code{corr_phase_incompleteness} can
be set at a time.}

\item{corr_phase_incompleteness}{Optional correlation between phase variation
(specifically how much the initial part of the time domain is expanded) and
the curve's incompleteness (specifically the length of the curve). This
parameter doesn't lead to the two measures having exactly the specified
correlation, but should be seen as a general parameter, higher (absolute)
values of which lead to a clearer association. Can have values between -1 and
\enumerate{
\item Defaults to 0. Only one of \code{corr_amplitude_incompleteness} /
\code{corr_phase_incompleteness} can be set at a time.
}}

\item{seed}{Optional numeric seed for \code{\link{set.seed}}}
}
\value{
If \code{is.null(FPCA_structure)}, then only the simulated dataset
is returned (including the underlying mean structure).
If \code{!is.null(FPCA_structure)}, then a list is returned
with both the simulated data and a dataset with information on the raw,
underlying FPCs.
}
\description{
Simulate potentially incomplete curves with some random warping and a
Generalized Functional Principal Component structure.
}
