% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStudy_algorithmWrappers.R
\name{varEM_wrapper}
\alias{varEM_wrapper}
\title{Wrapper to call varEM with assumed completeness}
\usage{
varEM_wrapper(
  data,
  job,
  instance,
  max_iterations,
  n_cores,
  distribution_reg = NULL
)
}
\arguments{
\item{data}{General arguments as requested by \code{batchtools}.}

\item{job}{General arguments as requested by \code{batchtools}.}

\item{instance}{A specific simulated data setting as returned by
\code{\link{simData_wrapper}}.}

\item{max_iterations}{Maximum number of iterations of the joint algorithm
of \code{register_fpca}.}

\item{n_cores}{Number of cores to use for computation.}

\item{distribution_reg}{Optional distribution (one of \code{c("gaussian","gamma")})
to be used in the registration step of \code{\link[registr]{register_fpca}}
instead of the original distribution of the simulated data found in
\code{instance$sim_settings$distribution}. When specified as \code{"t"} the
method is run with \code{"gaussian"} family.}
}
\value{
List with performance criteria.
}
\description{
Wrapper function to call varEM with Gaussian family and assumed completeness
in the simulation study. Written specifically for the use with the
\code{batchtools} package.
}
